Aggregate expression to tissue level

Aggregate expression to tissue level

Usage

aggregate_to_tissue(
  se,
  tissue_column,
  assay_name = NULL,
  aggregation_method = "median"
)

Arguments

se

A SummarizedExperiment object with expression data in TPM (or normalized by library size) and a colData slot.

tissue_column

Character indicating the name of the column in colData(se) containing tissue information for each sample.

assay_name

Character specifying the name of the assay to use. If NULL, the first assay will be extracted. Default: NULL.

aggregation_method

Character indicating the method to use to aggregate samples from the same tissue before calculating \(\tau\) (see details below). One of "median" or "mean". Default: "median".

Value

A gene expression matrix with genes in rows and tissues in columns.

Examples

# Load data
data(ath_se)
se <- ath_se

# Simulate different tissues
SummarizedExperiment::colData(se)$tissue <- c(rep("seed", 3), rep("leaf", 3))

# Aggregate expression
aexp <- aggregate_to_tissue(se, "tissue")