Calculate correlation coefficients and permutation P-values for gene pairs

Calculate correlation coefficients and permutation P-values for gene pairs

Usage

calculate_cor(se, dups, assay_name = NULL, method = "spearman")

Arguments

se

A SummarizedExperiment or SingleCellExperiment or SpatialExperiment object with expression data.

dups

A data frame of duplicate pairs, with duplicated genes for each pair in columns 1 and 2. Additional columns are allowed, and will be ignored if present.

assay_name

Character specifying the name of the assay to use. If NULL, the first assay will be extracted. Default: NULL.

method

Character indicating which correlation coefficient to compute. One of 'pearson', 'spearman', or 'kendall'. Default: 'spearman'.

Value

A data frame exactly as the one passed in dups, but with an extra column named cor (numeric) with correlation coefficients. Pairs will have NA in this column if i. at least of the genes in the pair is not present in the assay slot of se; or ii. at least of the genes in the pair has a standard deviation of 0.

Examples

# Load data
data(ath_se)
data(ath_dups)

# Subset data
dups <- ath_dups[ath_dups$type == "TD", ]
se <- ath_se

# Calculate correlation
cor_df <- calculate_cor(se, dups, method = "pearson")