Compare coexpression module eigengenes for diverged duplicate pairs

Compare coexpression module eigengenes for diverged duplicate pairs

Usage

compare_coex_me(module_comp, ME, cor_method = "spearman")

Arguments

module_comp

A data frame with module assignment comparison between duplicate pairs as returned by compare_coex_modules().

ME

A matrix or data frame of module eigengenes, with samples in rows and eigengenes in columns. This can be obtained, for instance, with BioNERO::exp2gcn().

cor_method

Character indicating the correlation method to use to calculate pairwise correlations between module eigengenes. Default: "spearman".

Value

A data frame with duplicate pairs (and any additional columns with metadata on pairs) and a column named ME_cor with correlation between eigengenes of the modules where each gene is.

Details

This function aims to derive a continuous measure from a binary variable (preserved vs diverged modules for genes in a duplicate pair). For that, this function calculates correlation coefficients between module eigengenes for diverged gene pairs. Such correlation coefficients can be interpreted as a measure of how much genes in a pair are diverged (i.e., if they are in very different modules, or in quite similar modules).

Examples

data(ath_dups)
data(genes_and_modules)
data(ME)
dups <- ath_dups
genes_modules <- genes_and_modules

# Compare
module_comp <- compare_coex_modules(dups, genes_and_modules)
comp <- compare_coex_me(module_comp, ME)